feat: M5 hardening - fixture suite, robustness tests, CI, packaging
R1 - "InnerTube response shapes change without notice" - has been the project's
most persistent risk since the analysis. It is now testable.

Fixtures: ten real responses captured by `cargo run --bin dump-fixtures` and
committed, with tests asserting what the parsers must extract - not merely that
they did not crash. At least 15 songs from a song search, 80% of them with a
duration and an artist, albums that are albums rather than tracks, an artist
page built from titled shelves, a queue of playable ids, and the presence of
playerOverlays, since that is where likeStatus lives.

Captures are anonymous: authenticated responses are full of personal data and
these files are committed. Tracking fields are stripped - trackingParams,
clickTrackingParams, visitorData, responseContext - which removes per-session
identifiers and a third of the bytes, and which no parser reads.

Verified the suite is not vacuous by reintroducing the exact M2 bug, reading
pageType by fixed path instead of searching for it: four tests fail.

Robustness: seeded mutation of the fixtures - dropping keys, nulling values,
swapping types, emptying arrays - asserting the parsers never panic and that
anything surviving is still internally consistent. Deterministic, so a failure
is reproducible from the seed. Degenerate and truncated documents covered too.
Real fuzz targets in fuzz/ for the response, LRC and key-binding parsers; they
need nightly, so CI runs the deterministic equivalent.

CI: fmt, clippy with -D warnings, build and test on Linux and macOS, plus a run
pinned to the 1.85 MSRV. A separate job runs the parser suites inside
`unshare -rn`, with no network at all: a red CI must mean our code broke, not
that an undocumented API changed overnight. Verified locally - 12 tests pass
with networking removed.

Packaging for AUR, Homebrew and cargo-binstall, plus a tagged release workflow
building x86-64 and aarch64 Linux and Apple Silicon. Both packages declare
ffmpeg and yt-dlp as runtime dependencies rather than suggestions, since an
install without them cannot decode or resolve anything.

The release profile keeps panic = "unwind" deliberately: the panic hook restores
the terminal, and abort would skip it.

Quality baseline: clippy clean across the workspace (two enums moved on every
keystroke had their Track payloads boxed), rustfmt clean, 53 tests, no build
warnings. LICENSE added, matching the GPL-3.0-or-later every manifest has
declared - though with no copyleft dependency in the tree, that remains a free
choice.

+++ crates/ytm-art/src/sixel.rs
@@ -51,7 +51,7 @@ pub fn encode(img: &RgbImage) -> String {
         if *u {
             let (r, g, b) = palette_rgb(i as u8);
             // Sixel colour components are percentages, not 0-255.
-            let pc = |v: u8| (v as u32 * 100 / 255) as u32;
+            let pc = |v: u8| v as u32 * 100 / 255;
             let _ = write!(out, "#{};2;{};{};{}", i, pc(r), pc(g), pc(b));
         }
     }
